refactor(entities): migrate entity DAL to entity homes, create 6 new entity folders
Moved entity-specific DAL from dal/server/<entity>/ to
entities/<entity>/dal/server/. Created new entity folders for:
accounts, activities, finance-options, media-files, push-subscriptions, users.

Each new entity gets a canonical constants.ts with its EntityName constant.
Activity added to ENTITY_NAMES (already had CASL rules) and removed from
the hand-maintained AppSubject union in types.ts.

Renamed api.ts → queries.ts in migrated DAL files to follow entity convention.
Fixed import sort order across ~20 files affected by path changes.
