+++ src/shared/entities/accounts/lib/constants.ts
@@ -1,0 +1,4 @@
+// Canonical entity-name constant for the Account entity. Source of truth
+// for `EntityName` and `AppSubject` (see domains/permissions/abilities.ts).
+// Note: Accounts are managed by better-auth — no EntityServerSpec.
+export const ACCOUNT = 'Account' as const
