+++ src/shared/entities/activities/lib/constants.ts
@@ -1,0 +1,3 @@
+// Canonical entity-name constant for the Activity entity. Source of truth
+// for `EntityName` and `AppSubject` (see domains/permissions/abilities.ts).
+export const ACTIVITY = 'Activity' as const
