+++ src/shared/entities/finance-options/lib/constants.ts
@@ -1,0 +1,3 @@
+// Canonical entity-name constant for the FinanceOption entity. Source of truth
+// for `EntityName` and `AppSubject` (see domains/permissions/abilities.ts).
+export const FINANCE_OPTION = 'FinanceOption' as const
