+++ src/shared/entities/push-subscriptions/lib/constants.ts
@@ -1,0 +1,4 @@
+// Canonical entity-name constant for the PushSubscription entity. Source of truth
+// for `EntityName` and `AppSubject` (see domains/permissions/abilities.ts).
+// Infrastructure entity — no EntityServerSpec.
+export const PUSH_SUBSCRIPTION = 'PushSubscription' as const
