+++ src/shared/entities/users/lib/constants.ts
@@ -1,0 +1,4 @@
+// Canonical entity-name constant for the User entity. Source of truth
+// for `EntityName` and `AppSubject` (see domains/permissions/abilities.ts).
+// Note: Users are managed by better-auth — no EntityServerSpec.
+export const USER = 'User' as const
