+++ src/trpc/routers/customers.router/business.router.ts
@@ -10,9 +10,6 @@ import z from 'zod'
 import env from '@/shared/config/server-env'
 import { intakeModes } from '@/shared/constants/enums'
 import { pipelines } from '@/shared/constants/enums/pipelines'
-import { deleteCustomer, getCustomer, getCustomers } from '@/shared/dal/server/customers/api'
-import { userCanSeeCustomer } from '@/shared/dal/server/customers/visibility'
-import { addParticipant } from '@/shared/dal/server/meetings/participants'
 import { buildFilterWhere } from '@/shared/dal/server/lib/query/filters'
 import { paginate } from '@/shared/dal/server/lib/query/output'
 import { dateRangeSchema, paginatedQueryInput } from '@/shared/dal/server/lib/query/schemas'
@@ -24,9 +21,12 @@ import { customerNotes } from '@/shared/db/schema/customer-notes'
 import { customers } from '@/shared/db/schema/customers'
 import { leadSourcesTable } from '@/shared/db/schema/lead-sources'
 import { meetings } from '@/shared/db/schema/meetings'
+import { deleteCustomer, getCustomer, getCustomers } from '@/shared/entities/customers/dal/server/queries'
+import { userCanSeeCustomer } from '@/shared/entities/customers/dal/server/visibility'
 import { derivedPipelineSql, derivedPipelineWhere } from '@/shared/entities/customers/lib/derived-pipeline-sql'
 import { gatedPhoneSql, hasSentProposalSql } from '@/shared/entities/customers/lib/phone-gating-sql'
 import { customerProfileSchema, financialProfileSchema, leadMetaSchema, propertyProfileSchema } from '@/shared/entities/customers/schemas'
+import { addParticipant } from '@/shared/entities/meetings/dal/server/participants'
 import { geocodeAddress } from '@/shared/services/providers/google-maps/geocode'
 
 import { createTRPCRouter } from '../../init'
