+++ src/trpc/routers/projects.router/media.router.ts
@@ -3,9 +3,9 @@ import { TRPCError } from '@trpc/server'
 import { eq, inArray } from 'drizzle-orm'
 import { z } from 'zod'
 import { mediaPhases } from '@/shared/constants/enums/media'
-import { resetOptimizationStatus } from '@/shared/dal/server/media-files/api'
 import { db } from '@/shared/db'
 import { insertMediaFilesSchema, mediaFiles } from '@/shared/db/schema'
+import { resetOptimizationStatus } from '@/shared/entities/media-files/dal/server/queries'
 import { R2_BUCKETS, R2_PUBLIC_DOMAINS } from '@/shared/services/providers/r2/buckets'
 import { getPresignedUploadUrl } from '@/shared/services/providers/r2/get-presigned-upload-url'
 import { deleteMediaWithVariants } from '@/shared/services/providers/r2/lib/delete-media-with-variants'
